UI development, created separate modules
the data browser now has a button to select a new set of samples, and more sample data is displayed. removed the weight input at the moment, since some data is missing this in the current dataset, age is now retrieved from the adnimerge csv which stores them as floating point numbers and not integers like in the dicom headers.