dataset pagination, state management
adjusted generator function in preprocessing so that the data browser can skip paths without generating from them, and generate individual samples without shuffling. added a class in interface to manage states and invoke windows, allowing pages to be closed without continuing the program or crashing, and also states can be transitioned between dynamically allowing for returning to dataset selection menu from the browser. care was taken to avoid recursion in this, as the browser must exit so that the volumes are deallocated from memory and don't stack when going to and from the selection screen. buttons added to the browser to pan through the dataset